Move LookupEntity out by itself
